requirements.txt
# Data Generation & Processing
faker==22.0.0
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2

# Date/Time handling
python-dateutil==2.8.2

# Data validation
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
jupyter==1.0.0
ipywidgets==8.1.1